Going back from Scala 2.9.1 to 2.8.1
There was a problem when using e.g. Option[Int] in Scala 2.9.1, which caused parameters using those types to not be parsed correctly as command line arguments. There was an ReviewStringException which clamed that the type was java.lang.Object, which I'm guessing suggest some type of inheritance problem. Moving back from Scala 2.8.1 solves the problem.