(Almost) working prototype
There is still a lot of checking of the parameters left to do, but the script basically seems to be working, with one exception. It breaks down when attempting to apply the variant recalibration to indels. I have as of yet not been able to figure out why that is.